commit LinkedListADT inteface along with updated MyLinkedList and updated MyLinkedListTest

+++ src/main/java/datastructures/interfaces/LinkedListADT.java
@@ -2,5 +2,5 @@ package datastructures.interfaces;
 
 public interface LinkedListADT<E> {
     void add(E element);
-    void remove(int index);
+    void remove(long index);
 }
